+++ src/conventional/bispectrumdx.py
@@ -138,7 +138,7 @@ def bispectrumdx(x, y, z, nfft=None, wind=None, nsamp=None, overlap=None):
   # accumulate triple products
   Bspec = np.zeros([nfft, nfft]) # the hankel mask (faster)
   mask = hankel(np.arange(nfft),np.array([nfft-1]+list(range(nfft-1))))
-  locseg = np.arange(nsamp).transpose()
+  locseg = np.arange(nsamp, dtype=int).transpose()
   x = x.ravel(order='F')
   y = y.ravel(order='F')
   z = z.ravel(order='F')
